+++ lib/drab/commander/config.ex
@@ -7,5 +7,7 @@ defmodule Drab.Commander.Config do
     ondisconnect: nil,
     # by default load Drab.Query and Drab.Modal
     modules: [Drab.Query, Drab.Modal],
-    access_session: []
+    access_session: [],
+    before_handler: [],
+    after_handler: []
 end