+++ test/drab/commander_test.exs
@@ -3,7 +3,8 @@ defmodule Drab.CommanderTest do
   doctest Drab.Commander
 
   defmodule TestCommander do
-    use Drab.Commander, onload: :onload_function, modules: [Drab.Query]
+    use Drab.Commander, modules: [Drab.Query]
+    onload :onload_function
   end
 
   test "__drab__/0 should return the valid config" do